+++ deltafi-ui/src/components/deletePolicy/DeletePolicyImportFile.vue
@@ -18,7 +18,7 @@
 
 <template>
   <span class="delete-policy-import-file">
-    <FileUpload ref="fileUploader" auto mode="basic" choose-label="Import Delete Policies" accept=".json,application/JSON" :file-limit="1" custom-upload class="p-button-sm p-button-secondary p-button-outlined mx-1" @uploader="preUploadValidation" @click="setOverlayPanelPosition" />
+    <FileUpload ref="fileUploader" auto mode="basic" choose-label="Import Policies" accept=".json,application/json" :file-limit="1" custom-upload class="p-button-sm p-button-secondary p-button-outlined mx-1" @uploader="preUploadValidation" @click="setOverlayPanelPosition" />
   </span>
   <OverlayPanel ref="errorOverlayPanel" dismissable show-close-icon @hide="clearUploadErrors">
     <Message severity="error" :sticky="true" class="mb-2 mt-0" :closable="false">