+++ deltafi-ui/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changes
+
+- Added System Snapshots page (#234)
+
 ## [0.98.3]
 
 ### Changed